feat: The Sidebar and header was added for the role of the patient and the psychologist
