+++ src/App.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import './App.css';
 
+import Home from './components/Home/Home';
 import Login from './components/Login/Login';
 import Dashboard from './components/Dashboard/Dashboard';
-import Home from './components/Home/Home';
 
 function App() {
 	return (
