--- src/components/Analysis/Analysis.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Analysis() {
-	return (
-		<div>
-			<h1>Hello Welcome to the Video Analysis</h1>
-		</div>
-	);
-}
-
-export default Analysis;
