--- src/components/Patients/Patients.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Patients() {
-	return (
-		<div>
-			<h1>Hello Welcome to the Patients</h1>
-		</div>
-	);
-}
-
-export default Patients;
